requirements.txt
Django~=3.1.6
selenium
django-allauth
django-environ
django-model-utils
shortuuid
django-extensions
django-crispy-forms==1.12.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-query-parameters

